+++ StudentSoup/src/main/frontend/src/components/board/BoardReviewFunction.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import './boardReviewFunction.scss';
 import Circle_human from 'assets/images/circle_human.png';
-import axiosInstance from 'apis/auth/AxiosInterceptor';
+import axiosInstance from 'apis/utils/AxiosInterceptor';
 import { useNavigate } from 'react-router-dom';
 import { Desktop, Mobile } from 'mediaQuery';
 import Swal from 'sweetalert2';
