+++ StudentSoup/src/main/frontend/src/components/common/MainNavbar.tsx
@@ -7,7 +7,7 @@ import { useEffect, useRef, useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faAngleRight, faBars, faXmark } from '@fortawesome/free-solid-svg-icons';
 import Swal from 'sweetalert2';
-import { postUserInfo } from 'apis/auth/BoardAPI';
+import { postUserInfo } from 'apis/api/BoardAPI';
 import { type userInformationType } from 'interfaces/BoardTypes';
 
 const MainNavbar = () => {
