+++ StudentSoup/src/main/frontend/src/pages/customerservice/CustomerService.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { postBoardCategory } from 'apis/auth/BoardAPI';
+import { postBoardCategory } from 'apis/api/BoardAPI';
 import { type BoardDataType } from 'interfaces/BoardTypes';
 import NoticeAndService from 'components/common/NoticeAndService';
 import Swal from 'sweetalert2';
